code cleanup: calibration details binning as int, not double
The binning value can be stored as an int, not a double.

This is backwards compatible with stored calibration values
as the stored values can be read back as integer values
even though they were written out as doubles.

+++ src/mount.cpp
@@ -1622,7 +1622,7 @@ void Mount::SaveCalibrationDetails(const CalibrationDetails& calDetails) const
         Debug.Write("Bogus guide speeds over-written in SaveCalibrationDetails\n");
     }
     pConfig->Profile.SetDouble(prefix + "ortho_error", calDetails.orthoError);
-    pConfig->Profile.SetDouble(prefix + "orig_binning", calDetails.origBinning);
+    pConfig->Profile.SetInt(prefix + "orig_binning", calDetails.origBinning);
     pConfig->Profile.SetString(prefix + "orig_timestamp", calDetails.origTimestamp);
     pConfig->Profile.SetInt(prefix + "orig_pierside", calDetails.origPierSide);
 
@@ -1789,7 +1789,7 @@ void Mount::LoadCalibrationDetails(CalibrationDetails *details) const
     details->orthoError = pConfig->Profile.GetDouble(prefix + "ortho_error", 0.0);
     details->raStepCount = pConfig->Profile.GetInt(prefix + "ra_step_count", 0);
     details->decStepCount = pConfig->Profile.GetInt(prefix + "dec_step_count", 0);
-    details->origBinning = pConfig->Profile.GetDouble(prefix + "orig_binning", 1.0);
+    details->origBinning = pConfig->Profile.GetInt(prefix + "orig_binning", 1);
     details->lastIssue = (CalibrationIssueType) pConfig->Profile.GetInt(prefix + "last_issue", 0);
     details->origTimestamp = pConfig->Profile.GetString(prefix + "orig_timestamp", "Unknown");
     details->origPierSide = pier_side(pConfig->Profile.GetInt(prefix + "orig_pierside", PIER_SIDE_UNKNOWN));
